+++ pysys-examples/internal/testcases/PySys_internal_053/run.py
@@ -6,7 +6,7 @@ class PySysTest(BaseTest):
 	def execute(self):
 		self.assertLineCount('file1.txt', filedir=self.input, expr='Foo', condition='==1', assertMessage='Looking for Foo')
 		self.assertLineCount('file1.txt', filedir=self.input, expr='Fi', condition='>=15', assertMessage='Looking for Fi')
-		self.waitForSignal('run.log', expr='Looking for Fi.*failed', timeout=5, ignores=[' DEBUG '])
+		self.waitForGrep('run.log', expr='Looking for Fi.*failed', timeout=5, ignores=[' DEBUG '])
 		
 		self.log.info('Copying run.log for later verification')
 		filecopy(os.path.join(self.output, 'run.log'), os.path.join(self.output, 'run.log.proc'))